song routes

+++ app/api/song_routes.py
@@ -1,8 +1,8 @@
-from flask import Blueprint, render_template, redirect, url_for
+from flask import Blueprint, render_template, redirect, url_for, jsonify
 from ..config import Config
 from flask_login import login_required, current_user
 from ..forms import CreateEditSongForm
-from ..models import db, Song, Like
+from ..models import db, Song, Like, Artist
 from flask_migrate import Migrate
 
 song_routes = Blueprint("songs", __name__)
@@ -11,20 +11,20 @@ song_routes = Blueprint("songs", __name__)
 @song_routes.route("/", methods=["GET"])
 def get_all_songs():
     songs = Song.query.all()
-    return {'songs': [song.to_dict() for song in songs]}
+    return jsonify({'songs': [song.to_dict() for song in songs]})
 
 # Get MY Songs - GET /api/songs/current
 @song_routes.route("/current", methods=["GET"])
 @login_required
 def get_my_songs():
     songs = Song.query.filter_by(user_id=current_user.id).all()
-    return {'songs': [song.to_dict() for song in songs]}
+    return jsonify({'songs': [song.to_dict() for song in songs]})
 
 # Get Song Details - GET /api/songs/:songId
 @song_routes.route("/<int:songId>", methods=["GET"])
 def get_song_details(songId):
     song = Song.query.filter_by(id=songId).one()
-    return song.to_dict()
+    return jsonify(song.to_dict())
 
 # @song_routes.route("/new", methods=["GET"])
 # @login_required
@@ -39,8 +39,14 @@ def create_song():
     form = CreateEditSongForm()
     if form.validate_on_submit():
         data = form.data
+        artist = Artist.query.filter_by(name=data['artist_name']).one()
+        if not artist:
+            new_artist = Artist(name=data['artist_name'])
+            db.session.add(new_artist)
+            db.session.commit()
+        artist = Artist.query.filter_by(name=data['artist_name']).one()
         new_song = Song(user_id=data['user_id'],
-                        artist_id=data['artist_id'],
+                        artist_id=artist.id,
                         title=data['title'],
                         lyrics=data['lyrics'],
                         url=data['url'],
@@ -48,7 +54,7 @@ def create_song():
                         release_date=data['release_date'])
         db.session.add(new_song)
         db.session.commit()
-        return redirect('/')
+        return jsonify(new_song.to_dict())
     if form.errors:
         return form.errors
 
@@ -63,7 +69,7 @@ def edit_song(songId):
     if form.validate_on_submit():
         data = form.data
         new_song = Song(user_id=data['user_id'],
-                        artist_id=data['artist_id'],
+                        artist_name=data['artist_name'],
                         title=data['title'],
                         lyrics=data['lyrics'],
                         url=data['url'],
@@ -71,7 +77,7 @@ def edit_song(songId):
                         release_date=data['release_date'])
         db.session.add(new_song)
         db.session.commit()
-        return redirect('/songId')
+        return jsonify(new_song.to_dict())
     if form.errors:
         return form.errors
 
@@ -80,6 +86,8 @@ def edit_song(songId):
 @login_required
 def delete_song(songId):
     song = Song.query.filter_by(id=songId).one()
+    if current_user.id != song.user_id:
+        return 'Forbidden'
     db.session.delete(song)
     db.session.commit()
-    return redirect('/')
+    return 'Song deleted'
